+++ aconfigure.ac
@@ -605,7 +605,7 @@ AC_ARG_ENABLE(upnp,
 		  AC_LINK_IFELSE([AC_LANG_PROGRAM([[#include <upnp/upnp.h>]],
 					          [UpnpInit2(NULL, 0);]
 					         )],
-		  		 [ CFLAGS="$CFLAGS -DPJLIB_UTIL_HAS_UPNP=1 $UPNP_CFLAGS"
+		  		 [ CFLAGS="$CFLAGS -DPJNATH_HAS_UPNP=1 $UPNP_CFLAGS"
 		  		   LDFLAGS="$LDFLAGS $UPNP_LDFLAGS $UPNP_LIBS"
 		  		   AC_MSG_RESULT(yes)
 		  		  ],
