+++ pjlib-util/include/pjlib-util.h
@@ -61,9 +61,6 @@
 /* Old STUN */
 #include <pjlib-util/stun_simple.h>
 
-/* UPnP */
-#include <pjlib-util/upnp.h>
-
 /* PCAP */
 #include <pjlib-util/pcap.h>
