+++ pjnath/include/pjnath/config.h
@@ -573,6 +573,19 @@
 						    PJ_VERSION_NUM_EXTRA)
 #endif
 
+
+/* **************************************************************************
+ * UPnP
+ */
+
+/* Default duration for searching UPnP Internet Gateway Devices (in seconds).
+ * Default: 5 seconds
+ */
+#ifndef PJ_UPNP_DEFAULT_SEARCH_TIME 
+#   define PJ_UPNP_DEFAULT_SEARCH_TIME 	5
+#endif
+
+
 /**
  * @}
  */
